eeprom: Provide easy-to-use overlays for hardware i2c-0 and i2c-1
There may be circumstances in which the user explicitly wants to use
the hardware I2C interfaces (such as when programming an EEPROM that
is not yet connected to the ID EEPROM pins).  Simplify this task by
allowing the I2C bus number to be specified as a constructor
parameter.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>